Move PackageScanner interface
